Add set_warm_start_from_previous_run method to all cpsat solvers
Originally when solving twice in a row with an ortools cpsat solver, it
starts again from scratch instead of continuing from the best solution it
found.

So we add this method (which was already in CpSatBinPackSolver) to warm
start the solver with the last solution found by itself.

We test it on coloring solver (and split the test_solvers.py module for
better readibility).
